Bug on cleaners reciever

+++ src/main/java/com/feerbox/client/registers/CleanerRegister.java
@@ -28,8 +28,10 @@ public class CleanerRegister extends Thread {
 					CleanerService.saveOrUpdate(cleaner);
 					companyId = cleaner.getCompany();
 				}
-				int removed = CleanerService.removeOtherCompanies(companyId);
-				logger.debug(removed+" cleaners removed from other companies than "+companyId);
+				if(companyId!=0){
+					int removed = CleanerService.removeOtherCompanies(companyId);
+					logger.debug(removed+" cleaners removed from other companies than "+companyId);
+				}
 			}
 			else{
 				logger.debug("No cleaners to update");
